+++ frontend/src/api/index.ts
@@ -53,6 +53,7 @@ export * as util from './util'
 export * as dataTracking from './dataTracking'
 export * as data from './data'
 export * as maps from './maps'
+export * as traffic from './traffic'
 
 export {default as client} from './client'
 export type * from './types'