+++ frontend/src/pages/helpcenter/station/manage/traffic.vue
@@ -1,0 +1,17 @@
+/*
+ *     SPDX-License-Identifier: AGPL-3.0-only
+ *
+ *     Copyright (C) RainbowDashLabs and Contributor
+ */
+<script setup lang="ts">
+import StationTrafficHelp from '~/views/helpcenter/stationview/manage/StationTrafficHelp'
+
+definePageMeta({
+  layout: 'helpcenter-station',
+  name: 'help-station-traffic',
+})
+</script>
+
+<template>
+  <StationTrafficHelp/>
+</template>