+++ frontend/src/pages/station/manage/traffic.vue
@@ -1,0 +1,17 @@
+/*
+ *     SPDX-License-Identifier: AGPL-3.0-only
+ *
+ *     Copyright (C) RainbowDashLabs and Contributor
+ */
+<script setup lang="ts">
+import ManageStationTrafficView from '~/views/stationview/manage/ManageStationTrafficView'
+
+definePageMeta({
+  layout: 'station',
+  name: 'station-traffic',
+})
+</script>
+
+<template>
+  <ManageStationTrafficView />
+</template>